[refactor] startsWith, isOutOf4ByteRange Util로 이동 (SRP)
