+++ TestShell/util.h
@@ -1,0 +1,5 @@
+#pragma once
+#include<string>
+
+bool startsWith(const std::string& str, const std::string& prefix);
+bool isOutOf4ByteRange(const std::string& hexStr);
